+++ src/cs-graphics/internal/tiny_gltf_helper.hpp
@@ -93,7 +93,7 @@ auto find_material_parameter(
   } else {
     auto const& parameter = it->second;
     T           value{};
-    for (int i = 0; i < std::min(size_t(value.length()), parameter.number_array.size()); ++i) {
+    for (size_t i = 0; i < std::min(size_t(value.length()), parameter.number_array.size()); ++i) {
       value[i] = static_cast<typename T::value_type>(parameter.number_array[i]);
     }
     return value;
